pyproject.toml
[project]
name = "1212"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "nump>=5.5.5.5",
]
